Added navbars to all html pages. Added profile page which shows pilot info.

+++ src/main/java/com/spring/fleetfindertest/model/Pilot.java
@@ -10,7 +10,7 @@ import javax.persistence.*;
 @Table(name = "characters")
 public class Pilot {
     @Id
-    private int charId;
+    private Long charId;
 
     @Column(name = "char_name")
     private String charName;
